requirements.txt
json
pyspark
tweepy==3.8.0
kafka-python==2.0.1
pysolr==3.8.1
configparser


#Reporting tool(>Python3.6)
superset==0.30.1
sqlalchemy==1.3.15
sqlalchemy-solr==0.1.dev0
sqlalchemy-utils==0.36.3